refactor: create astro-qiita-loader directory
